Split up classes in their own files
